+++ .vs/VSWorkspaceState.json
@@ -2,5 +2,6 @@
   "ExpandedNodes": [
     ""
   ],
+  "SelectedNode": "\\SickPersonaDies.sln",
   "PreviewInSolutionExplorer": false
 }
