+++ EndpointLambdas/EletronicPrice/EletronicPrice.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EletronicPrice", "EletronicPrice.csproj", "{19847534-E232-49AB-B91D-33B273EF7D75}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{19847534-E232-49AB-B91D-33B273EF7D75}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19847534-E232-49AB-B91D-33B273EF7D75}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19847534-E232-49AB-B91D-33B273EF7D75}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19847534-E232-49AB-B91D-33B273EF7D75}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6F43F62E-BA7F-47D9-9A22-A02931F3768C}
+	EndGlobalSection
+EndGlobal
