+++ EndpointLambdas/EletronicPrice/Properties/launchSettings.json
