+++ EndpointLambdas/FoodPrice/FoodPrice.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FoodPrice", "FoodPrice.csproj", "{495569F6-EE20-4D75-BB17-610056CC37D1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{495569F6-EE20-4D75-BB17-610056CC37D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{495569F6-EE20-4D75-BB17-610056CC37D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{495569F6-EE20-4D75-BB17-610056CC37D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{495569F6-EE20-4D75-BB17-610056CC37D1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D41E28F3-4C93-47E2-9DAB-AEA339CB2B64}
+	EndGlobalSection
+EndGlobal
