+++ EndpointLambdas/FoodPrice/Properties/launchSettings.json
