--- EndpointLambdas/GetEletronicPrice/GetEletronicPrice.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34723.18
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GetEletronicPrice", "GetEletronicPrice\GetEletronicPrice.csproj", "{1E73F283-E484-4B2B-A2BA-7AFF63DF817B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1E73F283-E484-4B2B-A2BA-7AFF63DF817B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1E73F283-E484-4B2B-A2BA-7AFF63DF817B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1E73F283-E484-4B2B-A2BA-7AFF63DF817B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1E73F283-E484-4B2B-A2BA-7AFF63DF817B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {14D4EC31-EAEA-4CD6-AAFB-5616B4C83267}
-	EndGlobalSection
-EndGlobal
