--- EndpointLambdas/GetFoodPrice/GetFoodPrice.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34723.18
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GetFoodPrice", "GetFoodPrice\GetFoodPrice.csproj", "{F825F621-FCA8-4125-A66A-67B36C9F36A5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F825F621-FCA8-4125-A66A-67B36C9F36A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F825F621-FCA8-4125-A66A-67B36C9F36A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F825F621-FCA8-4125-A66A-67B36C9F36A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F825F621-FCA8-4125-A66A-67B36C9F36A5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {96167FDD-AD94-4864-8794-17E227268DC2}
-	EndGlobalSection
-EndGlobal
