+++ EndpointLambdas/SalaryLambdas/Interfaces/IWageDetermination.cs
@@ -1,0 +1,14 @@
+﻿using SalaryLambdas.services;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SalaryLambdas.Interfaces
+{
+    public interface IWageDetermination
+    {
+        public Task<int> DetermineWageAsync();
+    }
+}
