+++ EndpointLambdas/SalaryLambdas/Interfaces/IWageService.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SalaryLambdas.Interfaces
+{
+    public interface IWageService
+    {
+        public Task<int> GetMinimumWageAsync();
+    }
+}
