+++ EndpointLambdas/SalaryLambdas/Models/PersonaWages.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SalaryLambdas.Models
+{
+    public class PersonaWages
+    {
+        public string personaId;
+        public int wage;
+    }
+}
