+++ EndpointLambdas/SalaryLambdas/SalaryLambdas.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SalaryLambdas", "SalaryLambdas.csproj", "{80BB4475-6EFE-47D2-991B-2B6C1DC86F2A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{80BB4475-6EFE-47D2-991B-2B6C1DC86F2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{80BB4475-6EFE-47D2-991B-2B6C1DC86F2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{80BB4475-6EFE-47D2-991B-2B6C1DC86F2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{80BB4475-6EFE-47D2-991B-2B6C1DC86F2A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {761EAF75-09E8-4AE0-A8D7-DA587C3F700E}
+	EndGlobalSection
+EndGlobal
