+++ EndpointLambdas/SickPersonaDies/SickPersonaDies.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SickPersonaDies", "SickPersonaDies.csproj", "{9A8354BC-77FD-410B-9141-8B006E338176}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9A8354BC-77FD-410B-9141-8B006E338176}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9A8354BC-77FD-410B-9141-8B006E338176}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9A8354BC-77FD-410B-9141-8B006E338176}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9A8354BC-77FD-410B-9141-8B006E338176}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D7E25A61-727A-4AF3-9719-48E39EADD666}
+	EndGlobalSection
+EndGlobal
