--- EndpointLambdas/salaryLambda/salaryLambda.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34723.18
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "salaryLambda", "salaryLambda\salaryLambda.csproj", "{C3A8BDD4-CEF4-4C7F-AC85-3FBCD3EBD08A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C3A8BDD4-CEF4-4C7F-AC85-3FBCD3EBD08A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C3A8BDD4-CEF4-4C7F-AC85-3FBCD3EBD08A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C3A8BDD4-CEF4-4C7F-AC85-3FBCD3EBD08A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C3A8BDD4-CEF4-4C7F-AC85-3FBCD3EBD08A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6F146529-E2B7-4238-B59D-42E6BB45A3E1}
-	EndGlobalSection
-EndGlobal
