--- EndpointLambdas/salaryLambda/salaryLambda/Interfaces/IWageDetermination.cs
@@ -1,14 +1,0 @@
-﻿using salaryLambda.services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace salaryLambda.Interfaces
-{
-    public interface IWageDetermination
-    {
-        public Task<int> DetermineWageAsync();
-    }
-}
