--- EndpointLambdas/salaryLambda/salaryLambda/Models/PersonaWages.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace salaryLambda.Models
-{
-    public class PersonaWages
-    {
-        public string personaId;
-        public int wage;
-    }
-}
