--- EndpointLambdas/salaryLambda/salaryLambda/aws-lambda-tools-defaults.json
@@ -1,16 +1,0 @@
-{
-  "Information": [
-    "This file provides default values for the deployment wizard inside Visual Studio and the AWS Lambda commands added to the .NET Core CLI.",
-    "To learn more about the Lambda commands with the .NET Core CLI execute the following command at the command line in the project root directory.",
-    "dotnet lambda help",
-    "All the command line options for the Lambda command can be specified in this file."
-  ],
-  "profile": "default",
-  "region": "eu-west-1",
-  "configuration": "Release",
-  "function-architecture": "x86_64",
-  "function-runtime": "dotnet8",
-  "function-memory-size": 512,
-  "function-timeout": 30,
-  "function-handler": "salaryLambda::salaryLambda.Function::FunctionHandler"
-}
